+++ tests/compile_fail/constructor/enum.rs
@@ -1,0 +1,6 @@
+#[derive(derive_more::Constructor)]
+enum Foo {
+    Bar(i32),
+}
+
+fn main() {}