+++ tests/compile_fail/constructor/unknown_attribute_argument.rs
@@ -1,0 +1,4 @@
+#[derive(derive_more::Constructor)]
+struct Foo(#[constructor(skip)] i32);
+
+fn main() {}